+++ DataGenerator/Data/Zones/ZoneInfoTables.cs
@@ -366,7 +366,7 @@ namespace DataGenerator.Data
             {
                 EffectTile exitTile = new EffectTile("stairs_go_down", true);
                 exitTile.TileStates.Set(new DestState(new SegLoc(moveBack, 1), true));
-                var step = new FloorStairsStep<MapGenContext, MapGenEntrance, MapGenExit>(new MapGenEntrance(Dir8.Down), new MapGenExit(exitTile));
+                var step = new FloorStairsStep<MapGenContext, MapGenEntrance, MapGenExit>(2, new MapGenEntrance(Dir8.Down), new MapGenExit(exitTile));
                 step.Filters.Add(new RoomFilterConnectivity(ConnectivityRoom.Connectivity.Main));
                 step.Filters.Add(new RoomFilterComponent(true, new BossRoom()));
                 layout.GenSteps.Add(PR_EXITS, step);